* map_drawing_show: (117 commits)
  moved websocket component to its own project
  no message
  code of maps and processing components migrated into addon libraries
  support for callback functions in processing component
  improved internatialization support
  fixed problem concerning widgetset compilation - method clone() cannot
have Override annotation due to gwt emulation
  internationalization support
  code cleaning and possible bug fixed
  update of web context
  ManagerUI draft
  fixed ImageSequenceTile
  updated classpath
  new: * implementation of wms layer, world coordinates * pan and zoom
control
  * renamed SRS constant to follow WMS 1.3 naming conventions * new
CRSUtility * use of proj4j library
  * redesigned Timer on client and server side as well. The server has
its own counting and is independent on client's counting events because
there were problems with losing client events somewhere between client
and server when using timer together with shortcut key actions. * fixed
css position parsing in MaskConnector * added synchronize clause into
ProcessEventManager (just in case) * added Video and Audio components
into slide xml parsing * added logging in some components * new
AbstractNonVisualComponentConnector * new AnnotationUtility
  * work on WMSLayer functionality
  * fixed translate transformation in vector graphics SVG implementation
  * renamed class VAbstractLayer to AbstractLayer * added class
InteractiveLayer to extend common behaviour of layers * redesigned
aligning of WMSTile in WMSLayer in single tile mode - it's done on
client side * added panning functionality to layers
  * temporarily removed Push mechanism - causes some difficulties
  * update to Vaadin 7.3.5
  ...

+++ src/com/tilioteo/hypothesis/event/ProcessEventType.java
@@ -8,15 +8,15 @@ package com.tilioteo.hypothesis.event;
  * 
  */
 public class ProcessEventType {
-	private int id;
+	private long id;
 	private String name;
 
-	public ProcessEventType(int id, String name) {
+	public ProcessEventType(long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
